Family page splits on litigation feature
