+++ src/components/pages/familyLitigationPage.tsx
@@ -1,25 +1,12 @@
-import { GetServerSideProps, InferGetStaticPropsType } from "next";
-
-import { ApiClient } from "@/api/http-common";
 import { Columns } from "@/components/atoms/columns/Columns";
 import Layout from "@/components/layouts/Main";
 import { ContentsSideBar } from "@/components/organisms/contentsSideBar/ContentsSideBar";
 import { FAMILY_PAGE_SIDE_BAR_ITEMS_SORTED } from "@/constants/sideBarItems";
-import { withEnvConfig } from "@/context/EnvConfig";
-import { TFamilyPage, TFeatureFlags, TTheme, TThemeConfig } from "@/types";
-import { getFeatureFlags } from "@/utils/featureFlags";
-import { isKnowledgeGraphEnabled } from "@/utils/features";
 import { getFamilyMetaDescription } from "@/utils/getFamilyMetaDescription";
-import { readConfigFile } from "@/utils/readConfigFile";
 
-interface IProps {
-  featureFlags: TFeatureFlags;
-  family: TFamilyPage;
-  theme: TTheme;
-  themeConfig: TThemeConfig;
-}
+import { IProps } from "./familyOriginalPage";
 
-const Family: InferGetStaticPropsType<typeof getServerSideProps> = ({ featureFlags, family, theme, themeConfig }: IProps) => {
+export const FamilyLitigationPage = ({ family, theme, themeConfig }: IProps) => {
   return (
     <Layout
       title={family.title}
@@ -47,37 +34,4 @@ const Family: InferGetStaticPropsType<typeof getServerSideProps> = ({ featureFla
       </Columns>
     </Layout>
   );
-};
-
-export default Family;
-
-export const getServerSideProps: GetServerSideProps = async (context) => {
-  const featureFlags = getFeatureFlags(context.req.cookies);
-
-  const id = context.params.id;
-  const client = new ApiClient(process.env.BACKEND_API_URL);
-
-  const theme = process.env.THEME;
-  const themeConfig = await readConfigFile(theme);
-
-  const knowledgeGraphEnabled = isKnowledgeGraphEnabled(featureFlags, themeConfig);
-  if (!knowledgeGraphEnabled) return { notFound: true };
-
-  let familyData: TFamilyPage;
-
-  try {
-    const { data: returnedData } = await client.get(`/documents/${id}`);
-    familyData = returnedData;
-  } catch (error) {
-    // TODO: handle error more elegantly
-  }
-
-  return {
-    props: withEnvConfig({
-      family: familyData,
-      theme,
-      themeConfig,
-      featureFlags,
-    }),
-  };
 };
